+++ system/admin/model/rolemodel_gen.go
@@ -3,113 +3,113 @@
 package model
 
 import (
-	"context"
-	"database/sql"
-	"fmt"
-	"strings"
-	"time"
+    "context"
+    "database/sql"
+    "fmt"
+    "strings"
+    "time"
 
-	"github.com/zeromicro/go-zero/core/stores/builder"
-	"github.com/zeromicro/go-zero/core/stores/cache"
-	"github.com/zeromicro/go-zero/core/stores/sqlc"
-	"github.com/zeromicro/go-zero/core/stores/sqlx"
-	"github.com/zeromicro/go-zero/core/stringx"
+    "github.com/zeromicro/go-zero/core/stores/builder"
+    "github.com/zeromicro/go-zero/core/stores/cache"
+    "github.com/zeromicro/go-zero/core/stores/sqlc"
+    "github.com/zeromicro/go-zero/core/stores/sqlx"
+    "github.com/zeromicro/go-zero/core/stringx"
 )
 
 var (
-	roleFieldNames          = builder.RawFieldNames(&Role{})
-	roleRows                = strings.Join(roleFieldNames, ",")
-	roleRowsExpectAutoSet   = strings.Join(stringx.Remove(roleFieldNames, "`create_time`", "`update_time`", "`create_at`", "`update_at`"), ",")
-	roleRowsWithPlaceHolder = strings.Join(stringx.Remove(roleFieldNames, "`id`", "`create_time`", "`update_time`", "`create_at`", "`update_at`"), "=?,") + "=?"
+    roleFieldNames          = builder.RawFieldNames(&Role{})
+    roleRows                = strings.Join(roleFieldNames, ",")
+    roleRowsExpectAutoSet   = strings.Join(stringx.Remove(roleFieldNames, "`create_time`", "`update_time`", "`create_at`", "`update_at`"), ",")
+    roleRowsWithPlaceHolder = strings.Join(stringx.Remove(roleFieldNames, "`id`", "`create_time`", "`update_time`", "`create_at`", "`update_at`"), "=?,") + "=?"
 
-	cacheRoleIdPrefix = "cache:role:id:"
+    cacheRoleIdPrefix = "cache:role:id:"
 )
 
 type (
-	roleModel interface {
-		Insert(ctx context.Context, data *Role) (sql.Result, error)
-		FindOne(ctx context.Context, id int64) (*Role, error)
-		Update(ctx context.Context, data *Role) error
-		Delete(ctx context.Context, id int64) error
-	}
+    roleModel interface {
+        Insert(ctx context.Context, data *Role) (sql.Result, error)
+        FindOne(ctx context.Context, id int64) (*Role, error)
+        Update(ctx context.Context, data *Role) error
+        Delete(ctx context.Context, id int64) error
+    }
 
-	defaultRoleModel struct {
-		sqlc.CachedConn
-		table string
-	}
+    defaultRoleModel struct {
+        sqlc.CachedConn
+        table string
+    }
 
-	Role struct {
-		Id         int64     `db:"id"`
-		ParentId   int64     `db:"parent_id"` // 父级id   默认0为顶级
-		Name       string    `db:"name"`
-		Info       string    `db:"info"`
-		Code       string    `db:"code"`
-		Status     int64     `db:"status"`    // 状态 0=开启 1=关闭
-		IsDelete   int64     `db:"is_delete"` // 是否删 0=否 1=是
-		CreateTime time.Time `db:"create_time"`
-	}
+    Role struct {
+        Id         uint       `db:"id"`
+        ParentId   uint       `db:"parent_id"` // 父级id   默认0为顶级
+        Name       string    `db:"name"`
+        Info       string    `db:"info"`
+        Code       string    `db:"code"`
+        Status     uint8       `db:"status"`    // 状态 0=开启 1=关闭
+        IsDelete   uint8       `db:"is_delete"` // 是否删 0=否 1=是
+        CreateTime time.Time `db:"create_time"`
+    }
 )
 
 func newRoleModel(conn sqlx.SqlConn, c cache.CacheConf) *defaultRoleModel {
-	return &defaultRoleModel{
-		CachedConn: sqlc.NewConn(conn, c),
-		table:      "`role`",
-	}
+    return &defaultRoleModel{
+        CachedConn: sqlc.NewConn(conn, c),
+        table:      "`role`",
+    }
 }
 
 func (m *defaultRoleModel) Delete(ctx context.Context, id int64) error {
-	roleIdKey := fmt.Sprintf("%s%v", cacheRoleIdPrefix, id)
-	_, err := m.ExecCtx(ctx, func(ctx context.Context, conn sqlx.SqlConn) (result sql.Result, err error) {
-		query := fmt.Sprintf("delete from %s where `id` = ?", m.table)
-		return conn.ExecCtx(ctx, query, id)
-	}, roleIdKey)
-	return err
+    roleIdKey := fmt.Sprintf("%s%v", cacheRoleIdPrefix, id)
+    _, err := m.ExecCtx(ctx, func(ctx context.Context, conn sqlx.SqlConn) (result sql.Result, err error) {
+        query := fmt.Sprintf("delete from %s where `id` = ?", m.table)
+        return conn.ExecCtx(ctx, query, id)
+    }, roleIdKey)
+    return err
 }
 
 func (m *defaultRoleModel) FindOne(ctx context.Context, id int64) (*Role, error) {
-	roleIdKey := fmt.Sprintf("%s%v", cacheRoleIdPrefix, id)
-	var resp Role
-	err := m.QueryRowCtx(ctx, &resp, roleIdKey, func(ctx context.Context, conn sqlx.SqlConn, v interface{}) error {
-		query := fmt.Sprintf("select %s from %s where `id` = ? limit 1", roleRows, m.table)
-		return conn.QueryRowCtx(ctx, v, query, id)
-	})
-	switch err {
-	case nil:
-		return &resp, nil
-	case sqlc.ErrNotFound:
-		return nil, ErrNotFound
-	default:
-		return nil, err
-	}
+    roleIdKey := fmt.Sprintf("%s%v", cacheRoleIdPrefix, id)
+    var resp Role
+    err := m.QueryRowCtx(ctx, &resp, roleIdKey, func(ctx context.Context, conn sqlx.SqlConn, v interface{}) error {
+        query := fmt.Sprintf("select %s from %s where `id` = ? limit 1", roleRows, m.table)
+        return conn.QueryRowCtx(ctx, v, query, id)
+    })
+    switch err {
+    case nil:
+        return &resp, nil
+    case sqlc.ErrNotFound:
+        return nil, ErrNotFound
+    default:
+        return nil, err
+    }
 }
 
 func (m *defaultRoleModel) Insert(ctx context.Context, data *Role) (sql.Result, error) {
-	roleIdKey := fmt.Sprintf("%s%v", cacheRoleIdPrefix, data.Id)
-	ret, err := m.ExecCtx(ctx, func(ctx context.Context, conn sqlx.SqlConn) (result sql.Result, err error) {
-		query := fmt.Sprintf("insert into %s (%s) values (?, ?, ?, ?, ?, ?, ?)", m.table, roleRowsExpectAutoSet)
-		return conn.ExecCtx(ctx, query, data.Id, data.ParentId, data.Name, data.Info, data.Code, data.Status, data.IsDelete)
-	}, roleIdKey)
-	return ret, err
+    roleIdKey := fmt.Sprintf("%s%v", cacheRoleIdPrefix, data.Id)
+    ret, err := m.ExecCtx(ctx, func(ctx context.Context, conn sqlx.SqlConn) (result sql.Result, err error) {
+        query := fmt.Sprintf("insert into %s (%s) values (?, ?, ?, ?, ?, ?, ?)", m.table, roleRowsExpectAutoSet)
+        return conn.ExecCtx(ctx, query, data.Id, data.ParentId, data.Name, data.Info, data.Code, data.Status, data.IsDelete)
+    }, roleIdKey)
+    return ret, err
 }
 
 func (m *defaultRoleModel) Update(ctx context.Context, data *Role) error {
-	roleIdKey := fmt.Sprintf("%s%v", cacheRoleIdPrefix, data.Id)
-	_, err := m.ExecCtx(ctx, func(ctx context.Context, conn sqlx.SqlConn) (result sql.Result, err error) {
-		query := fmt.Sprintf("update %s set %s where `id` = ?", m.table, roleRowsWithPlaceHolder)
-		return conn.ExecCtx(ctx, query, data.ParentId, data.Name, data.Info, data.Code, data.Status, data.IsDelete, data.Id)
-	}, roleIdKey)
-	return err
+    roleIdKey := fmt.Sprintf("%s%v", cacheRoleIdPrefix, data.Id)
+    _, err := m.ExecCtx(ctx, func(ctx context.Context, conn sqlx.SqlConn) (result sql.Result, err error) {
+        query := fmt.Sprintf("update %s set %s where `id` = ?", m.table, roleRowsWithPlaceHolder)
+        return conn.ExecCtx(ctx, query, data.ParentId, data.Name, data.Info, data.Code, data.Status, data.IsDelete, data.Id)
+    }, roleIdKey)
+    return err
 }
 
 func (m *defaultRoleModel) formatPrimary(primary interface{}) string {
-	return fmt.Sprintf("%s%v", cacheRoleIdPrefix, primary)
+    return fmt.Sprintf("%s%v", cacheRoleIdPrefix, primary)
 }
 
 func (m *defaultRoleModel) queryPrimary(ctx context.Context, conn sqlx.SqlConn, v, primary interface{}) error {
-	query := fmt.Sprintf("select %s from %s where `id` = ? limit 1", roleRows, m.table)
-	return conn.QueryRowCtx(ctx, v, query, primary)
+    query := fmt.Sprintf("select %s from %s where `id` = ? limit 1", roleRows, m.table)
+    return conn.QueryRowCtx(ctx, v, query, primary)
 }
 
 func (m *defaultRoleModel) tableName() string {
-	return m.table
+    return m.table
 }
